News-Website/requirement.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.5
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-flatpickr==1.0.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-markdown-deux==1.0.5
django-mptt==0.11.0
django-taggit==1.2.0
django-taggit-serializer==0.1.7
djangorestframework==3.12.4
djangotoolbox==1.8.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==4.0.1
Markdown==3.2.1
markdown2==2.4.0
mysql-connector-python==8.0.23
Pillow==7.2.0
protobuf==3.15.8
psycopg2==2.8.6
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pyserial==3.5
python-decouple==3.4
pytz==2020.1
requests==2.25.1
schedule==1.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.26.4
values==2020.12.3
whitenoise==5.2.0
zipp==3.4.1
